Project/frontend/requirements.txt
streamlit>=1.29.0
pandas>=2.2.2
numpy>=1.26.4
plotly==5.17.0
jieba==0.42.1
requests==2.31.0
vaderSentiment==3.3.2
pillow==10.1.0
openpyxl>=3.1.0
wordcloud==1.9.2
seaborn==0.13.0
matplotlib==3.8.2
dashscope>=1.14.0
supabase>=2.6.0